plan, ... docstrings - point to the full docs
When a new user comes upon next.jdbc, she is most likely to check the docstring of the functions being used. She might however not be aware of the Getting Started guide and that examples and full explanation are there. So I think it will be very helpful for such people to point them to the docs, from the docstring [of the most used functions].