Extend ReadableColumn protocol with more rsmeta
There is a use case for needing access to ResultSetMetaData inside
read-column-by-label.

SQLite doesn't have a boolean affinity. It stores booleans as integers 0
or 1, so that's what the JDBC driver gives us, and without access to
ResultSetMetaData we have no way of knowing the intended affinity of the column
that the value came from. But by looking at `getColumnTypeName` on the
ResultSetMetaData and seeing if it's "BOOL", we can perform the
conversion in the ReadableColumn implementations.